requirements.txt
appdirs==1.4.3
autopep8==1.3.1
cycler==0.10.0
matplotlib==2.0.1
mss==2.0.22
numpy==1.12.1
oauthlib==2.0.2
olefile==0.44
opencv-python==3.2.0.7
packaging==16.8
Pillow==4.1.1
PyAutoGUI==0.9.36
pycodestyle==2.3.1
PyMsgBox==1.0.6
pyparsing==2.2.0
pyperclip==1.5.27
PyScreeze==0.1.9
python-dateutil==2.6.0
python-xlib==0.19
PyTweening==1.0.3
pytz==2017.2
PyUserInput==0.1.11
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
tweepy==3.5.0
